Other iterations/video-analytics-system/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
pillow==10.2.0
websockets==12.0
insightface==0.7.3
onnxruntime==1.16.3
